+++ src/main/kotlin/com/procurement/notice/model/ocds/CoefficientValue.kt
@@ -6,5 +6,5 @@ sealed class CoefficientValue {
     data class AsBoolean(val value: Boolean) : CoefficientValue()
     data class AsString(val value: String) : CoefficientValue()
     data class AsNumber(val value: BigDecimal) : CoefficientValue()
-    data class AsInteger(val value: Int) : CoefficientValue()
+    data class AsInteger(val value: Long) : CoefficientValue()
 }
